Remove poetry lock (#399)
Reasoning: 

For some reason GitHub environments do not like the latest version of
`poetry` (v1.5.1 as of now). The only version that works without
problems is v1.3.2. Hence `qc.yml` is pinned to that version. By adding
`poetry.lock` to `.gitignore`, GitHub generates its own `poetry.lock` in
its environment and there won't be errors while installing dependencies
in the GitHub environment.

This is the solution proposed via this PR.