Use `curies.Converter` in OntoPortal exporter
This PR begins to switch the OntoPortal exporter to use `curies.Converter` for expansion and compression. Notably, this replaces ad hoc code with battle tested and provably correct expansion code.

However, it's not so obvious what the data is supposed to look like in the end since the test cases only check that there is some data, and not actually what it looks like.

@syphax-bouazzouni can you help build up the test cases so we can be more sure that this is working correctly? Is there documentation somewhere for the schema of JSON that this exporter corresponds to?