Add converter to MappingSetDocument and MappingSetDataFrame (#417)
This PR switches the MappingSetDocument to use a `curies.Converter`
instead of a simple prefix map. It exposes `prefix_map` as a property to
maintain backward compatibility.

Conversely, it exposes a `converter` property on the MappingSetDataFrame
for forwards compatibility.

It then updates several utilities to take advantage of these data
structures when converting between objects. It also fixes two bugs when
generating cliques: 1) the existing document's license is now retained
and 2) each clique is given a different unique ID