Implement custom validation on top of LinkML.
The SSSOM schema will soon no longer define the `subject_id` slot as
being inconditionally required. Instead, it will only be required iff
`subject_type` is not set to `rdfs literal`; if `subject_type` is set to
`rdfs literal`, it is `subject_label` that will be required instead.

Unfortunately, even though this variable constraint is formally
expressed in the SSSOM schema, the generated Python constructor does not
enforce it, and simply treat `subject_id` as a non-required slot.

This means that we can no longer rely on the Mapping constructor to
correctly flag mappings with a missing `subject_id`. Instead, we need to
perform a post-constructor validation by manually looking at
`subject_id` and `subject_type`. This is what we do here.

(All the above also applies, mutatis mutandis, to `object_id`.)