Fix reordering of columns after inversion.
When inverting a data frame, the reordering of columns post-inversion
should not rely on the columns that existed in the original data frame
(pre-inversion), because it is not guaranted that all columns in the
original data frame will exist in the inverted data frame: if the
original frame contained, e.g., a `subject_label` column but did not
contain a `object_label` column, then the inverted frame will contain an
`object_label` column but will _not_ contain a `subject_label` column.

Instead, we use the dedicated sort_df_rows_columns function to perform
the reordering. We sort the columns only, as there should be no need at
this stage to sort the rows.