Add test for inversion of an "asymmetric" mapping set.
We add a new test file containing an example of an "asymmetric" or
"imbalanced" set: that is, a set where not all "subject_*" columns have
an equivalent "object_*" column and vice versa.

The set is derived from the "basic.tsv" set by: (1) removing most of the
set metadata (not needed for the test); (2) removing most of the rows,
leaving just a handful (more than enough); (3) removing most of the
columns. In the columns we keep, we specifically keep "subject_label"
but NOT "object_label" (thereby creating the "asymmetry"), and
conversely we keep "object_source" but NOT "subject_source" (another
asymmetry).

We then add a new test case in tests/test_utils.py in which we try
inverting the asymmetric test (checking that it does not trigger the
issue #554) and check that the inverted object labels correspond to the
original subject labels.