Fix ChainedAssignmentError in from_sssom_dataframe
Warning: ChainedAssignmentError: A value is trying to be set on a copy of a DataFrame or Series through chained assignment using an inplace method.
When using the Copy-on-Write mode, such inplace method never works to update the original DataFrame or Series, because the intermediate object on which we are setting values always behaves as a copy.

I just to just not use "inplace" here and the warning does not happen.