Refactor df.infer_object call to account for diverging pandas versions
To ensure the old behaviour is retained, @gouttegd convinced me to not be lazy and actually add the "df.infer_objects" functionality when replacing empty values with nan's. Should github still exist when you read this, you can find his feedback here: https://github.com/mapping-commons/sssom-py/pull/561/files#r1835148927

As he questions, it is not clear this is _strictly needed_, but its probably safer to so if we dont know if its needed.