Extend parse_sssom_table to report wrong prefixes and metadata
This also comes with the introduction of a strict mode, which allows draconian failing in cases where small violations are encountered (by default, we drop unknown metadata elements, and redefined built in prefixes).

So the idea is:

Look for some basic errors before the actual parsing occurs, fail if in strict mode, if not proceed to normal mode with repair.