Add missing slots to COLUMN_INVERT_DICTIONARY.
Mapping inversion requires, among other things, to know which are the
slots whose value must be inverted. For example, knowing that
`subject_id` must be inverted with `object_id`, `subject_label` must be
inverted with `object_label`, and so on.

This knowledge is provided, in SSSOM-Py, in the
`COLUMN_INVERT_DICTIONARY` map in the `sssom.constants` module.

That map is missing the `subject_type` and `object_type` slots,
resulting in SSSOM-Py being unable to invert any mapping set that
contains one of those slots.

We add the missing entries to the `COLUMN_INVERT_DICTIONARY` map.

closes #554