Add all "asymmetric" slots to the asymmetric test case.
The test supposed to ensure that "asymmetric" mapping sets (sets that
contain some `subject_*` columns without their `object_*` counterparts,
or the other way round) only contains a couple of asymmetric slots.

As a result, it didn't catch the problem caused by the absence of
`subject_type` and `object_type` in the COLUMN_INVERT_DICTIONARY map.

We fill the `asymmetric.tsv` set so that it contains _all_ the slots
that require switching when a mapping set is inverted.