Do not load scipy unconditionally.
SciPy is considered an optional dependency, so the sssom.cli module
should not unconditionally attempt to load it -- this makes scipy a de
facto mandatory dependency.

SciPy is only needed for the `correlations` subcommand, so when that
command is used we try to load the scipy module, and error out cleanly
with a message indicating that this particular subcommand requires
SciPy.