Error out gracefully on missing optional module (#600)
* Do not load scipy unconditionally.

SciPy is considered an optional dependency, so the sssom.cli module
should not unconditionally attempt to load it -- this makes scipy a de
facto mandatory dependency.

SciPy is only needed for the `correlations` subcommand, so when that
command is used we try to load the scipy module, and error out cleanly
with a message indicating that this particular subcommand requires
SciPy.

* Error out gracefully when an optional module is missing.

Apply to the optional modules `networkx` and `pansql` the same principle
as for the `scipy` module: when they are not available, the commands
that depend on them should fail gracefully with an explicit error
message, rather than spit out an uncaught ModuleNotFoundError.

* Apply black-mandated style.