Add support for propagating/condensing slots.
Add two new methods to the MappingSetDataFrame class: `propagate` and
`condense`.

`propagate` implements propagation of propagatable slots: when such a
slot is present on the MappingSet but _not_ in individual records (i.e.
the data frame has no column for that slot), the value is "propagated"
down to the individual records. That is, a column for that slot is added
to the data frame, with the same value for all records, while the
original slot is removed from the set metadata.

`condense` performs the opposite operation: when a propagatable slot has
only a single value across all records, the value is set once and for
all in the set metadata and removed from the individual records (the
entire column is removed from the data frame).