Only condense if the slot has no value at set-level.
As par the SSSOM specification, condensing a slot is only allowed if (1)
all records have the same value for that slot *and* (2) the mapping set
either has no value for that slot, or already has the same value as the
value in the mapping records.

This commit adds the check needed to ensure that condition (2) is met.