Be more verbose about propagation and condensation.
Provide some explanation in the docstrings for the `propagate` and
`condense` methods, to remind people of what propagation and
condensation are and the conditions under which these operations are
allowed as per the SSSOM specification.

Also emit a warning when (1) a slot cannot be propagated because
individual records already have a conflicting value, or (2) a slot
cannot be condensed because the mapping set already has a conflicting
value.