Add fill_empty option to the propagate method.
Update the MappingSetDataFrame#propagate() method so that it accepts an
optional fill_empty argument.

When fill_empty is True, the method deviates from the behaviour mandated
by the SSSOM specification in that it will propagate a slot to all
records that do not have a value for that slot, even if other records do
have a value (the spec-mandated behaviour is that, as soon as some
records have a value for the slot, then the slot must be considered as
non-propagatable).