Optimize split_dataframe_by_prefix
There was a bit too much iteration going on in
`parsers.split_dataframe_by_prefix`-- for every combination of
prefixes*relations*objects, the entire input dataframe was being
iterated over. This was a lot of overhead, when it only needs to be gone
through once.

This is basically the same approach used in sssom-java (see note at
bottom), but there is some more work being done to keep logging
consistent with the previous approach.

In testing splitting `mondo.sssom.tsv` (a 25mb file), the time to split
went from ~5 minutes to ~50 seconds. Nearly all of the time spent is on
calling `parsers.from_sssom_dataframe`, which is slow but inevitable.

<https://github.com/gouttegd/sssom-java/blob/main/cli/src/main/java/org/incenp/obofoundry/sssom/cli/SimpleCLI.java#L646-L670>