Use `parse_curie` when grouping mappings; clarify grouping process
Renamed the NamedTuple which dictates how mappings should be split, from
SSSOMSplitTriple to SSSOMSplitGroup. This tuple consists of the
combination of:

  1. subject_prefix
  2. object_prefix
  3. relation_curie

This clarifies the purpose of the class-- it may be useful to group
instead by only 1 & 2 (this is what the `split` cli option in sssom-java
does).

Added/clarified documentation.