Add helper function to parse SSSOM version number.
Add a small helper function to turn a "X.Y" string into a valid SSSOM
version number represented as a tuple of integers (X, Y).

Instead of working on the input string directly (splitting into two
substrings, then converting the substrings to integers), we first
convert the string into a SssomVersionEnum object, from which we get the
string back. This is so we can rely on the LinkML-generated code to
automatically check that the provided value is a valid value that
correctly identifies a valid SSSOM version, without having to embed into
the method the knowledge of which versions are valid at any given time.