Add method to enforce compliance with a given SSSOM version.
Add a new method `MappingSetDataFrame#enforce_compliance()` to ensure
that a mapping set is compliant with a given version of the SSSOM
specification, by removing any slot or slot value that has only been
defined in a later version.

The method can also be used to optionally remove any extra non-standard
slot that has not been properly declared as an extension slot
(strict=True).