Centralize knowledge about newly added enum values.
Currently the fact that a given enum value has been added in a specific
version of the specification (for example, "composed entity expression"
is new in 1.1) is not formally recorded anywhere, and wherever that
information is needed we need custom code to deal with it.

This commit adds a new constant `NEW_ENUM_VALUES` that provides that
information once and for all.

This makes the code for both the `get_compatible_version()` and
`enforce_version()` much simpler, avoids duplicated information, and
will make it easier to later cope with future new enum values.