Fix support for literal mappings. (#624)
Since the LinkML-generated Mapping class does not, in itself, validates
the post-conditions specified in the schema, it cannot validate that a
Mapping object is created with either a subject_id or (if subject_type
is `rdfs literal`) a subject_label (likewise for the object side).

Instead, that validation is performed by custom code executed as part of
the parsing code.

However, the present validation is bogus because the check on the value
of `subject_type` and `object_type` is trying to compare the value of
the slot to a literal string, instead of a EntityTypeEnum object. We fix
that comparison here.