Encapsulate new enum logic into the SSSOMSchemaView class.
Add a `get_new_enum_values()` method to the SSSOMSchemaView class to get
enum values that were introduced after a given version of the spec.

First, this dispenses client code from having to explicitly import the
NEW_ENUM_VALUES constant. Second, hopefully in the future we might be
able to get the information about new enum values directly from the
LinkML schema rather than from a hard-coded list, and when that happens
we will simply have to update the `get_new_enum_values()` method without
impacting the code that is calling that method.