Use SSSOMSchemaView() when needing to get the list of mapping slots.
The `inject_metadata_into_df` function (whose entire purpose is very
dubious but that is another matter) needs a list of the slots available
in the mapping class. It gets that list by doing its own loading of the
SSSOM schema. It does so without specifying any encoding for the schema
file, and apparently the default YAML parser, when not provided with an
explicit encoding on Windows, assumes the encoding is CP-1252. This
results in a failure to load the schema if it happens to contain
UTF-8-encoded non-ASCII characters (which in version 1.1.0a2 it does,
because we are no longer in the 1980s).

This commit updates the `inject_metadata_into_df` function so that it
reuses the SSSOMSchemaView class, which is precisely intended to give
access to the contents of the schema.