Add method to enforce compliance with a given version of the SSSOM specification. (#616)
This commit adds two helper methods to the `MappingSetDataFrame`
class.

The `get_compatible_version()` method is intended to automatically
determine the minimum version of the SSSOM specification the set is
compatible with -- that is, the earliest version that defines all the
slots and all the enum values present in the set.

The `enforce_compliance()` method is intended to ensure that a
mapping set is compliant with a given version of the SSSOM
specification, by removing any slot or slot value that has only been
defined in a later version.

The method can also be used to optionally remove any extra non-standard
slot that has not been properly declared as an extension slot
(strict=True).