Fix msdf_merge util function. (#627)
The `msdf_merge` function attempts to "propagate" slots before merging
the sets, but is doing so without any regard for which slots should
actually be propagated.

In addition, it attempts to inject in every individual mapping a
`mapping_set_source` slot pointing to the ID of the original set that
contained the mapping, but this is invalid as there is _no_
`mapping_set_source` slot on indivdual mapping records -- the slot
intended to capture the set from which a record came from is
`mapping_source`.

Lastly, the function also attempts to drop duplicates after the sets
have been merged, but the detection of duplicates is prevented by (1)
the incorrect propagation of non-propagatable slots (which can cause two
otherwise identical records in two different sets to appear different,
if the metadata of the sets contain different wrongly propagated slots),
and (2) the injection of the `mapping_set_source` slot.

This commit fixes all those issues by deleting the bogus
`inject_metadata_into_df` function and replacing by a call to
`msdf.propagate()`, which implements propagation correctly. It then
manually inject the correct `mapping_source` slot if possible, and if so
ignore the injected slot when attempting to drop duplicates.