Enable automatic propagation and condensation.
Add a `propagate` parameter to all `parse_*` functions. When that
parameter is true, the parsing function will automatically propagate all
condensed slots in the parsed mapping set, prior to returning a
MappingSetDataFrame.

For the `parse_obographs_json` function, the parameter does not actually
make much sense (a mapping set extracted from a OBOGraph-JSON document
cannot be in a condensed state), but having all parsing functions accept
the same `propagate` parameter allows to keep using the
`get_parsing_function` logic without having to make a special case for
the functions that would not accept a `propagate` parameter. Likewise
for the `parse_alignment_xml` function.

The `propagate` parameter defaults to True -- this is the behaviour
recommended by the SSSOM specification --, except for the aforementioned
`parse_obographs_json` and `parse_alignment_xml` functions.

Propagation is normally performed by calling the
`MappingSetDataFrame#propagate()` method, but in the case of the
`parse_sssom_table()` function this cannot work, because propagation
needs to be performed before we can get a MappingSetDataFrame instance
(otherwise we could fail to construct a MappingSetDataFrame if the
mapping set contains literal mappings and the `subject_type` or
`object_type` slot is condensed). So we use a new
`propagate_condensed_slots()` method instead, which does not require a
MappingSetDataFrame object.

Conversely, we also add a `condense` parameter to all the `write_*`
functions, to perform the opposite operation prior to writing a set. The
parameter defaults to True for all writing functions, except `write_rdf`
since writing condensed slots in RDF is most likely not wanted.