Update tests for the new behaviour of parsing functions.
Now that (most) parsing functions default to propagate condensed slots,
some tests need to be updated. Parsing-time propagation must be disabled
for:

* some tests that are performed on "half-condensed" test files and do
  not expect the structure of the set to be modified;
* the tests for the actual propagation/condensation feature (can't
  properly test propagation if the parser already did it for us).

In addition, we add a test to check that parsing-time condensation does
allow to successfully read a set containing literal mappings when the
object_type slot is condensed (#606).