Add --[no-]propagate and --[no-]condense options.
Now that parsing and writing functions default to automatically
propagate and condense (respectively) sets, we add `--no-propagate`
and `--no-condense` options to some commands to alter the default
behaviour.

Specifically, we add such options to:

* `sssom convert`,
* `sssom parse`,
* `sssom validate`,
* and `sssom merge`.

It is currently not deemed useful (by me at least) to add those options
to every single command.