Slightly overhaul the value converters.
Introduce a `CurieConverterProvider` interface so that we can construct
all value converters even before we get the CURIE converter that the
EntityReferenceValueConverter will need to convert EntityReference-typed
value.

This is the first step towards making it possible to create only one
instance of the MappingSetRDFConverter class (maybe even making it a
Singleton) that can be used to serialize/deserialize many different MSDS
objects -- currently, the class can only really be used once, because it
expects to get the CURIE converter at construction time, and in most
cases the CURIE converter will be specific to only one MSDF or one RDF
graph.